Task 3 - Calculator
The task consisted of creating a Calculator using HTML, CSS and JAVASCRIPT , it involved designing an interactive interface with buttons for essential operations like addition, subtraction, multiplication, and division. I implemented a display screen to showcase user inputs and results, making use of CSS grid system for button alignments.

The project required the application of event listeners, if-else statements, operators, and loops to effectively handle user interactions and perform accurate calculations. This experience not only enhanced my skills but also demonstrated that even with a foundational understanding of these technologies, I could deliver a practical and functional solution.